+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+## Added
+- Added the `--skip-hosts` option to th `apply` command to skip editing the hosts file. ([#110](https://github.com/craftcms/nitro/issues/110))
+
 ## Changed
 - The `destroy` command now prompts for verification before machine destruction. ([#116](https://github.com/craftcms/nitro/issues/116))
 - The `self-update` command no longer prompts for installation upon completion. ([#115](https://github.com/craftcms/nitro/issues/115))